CDCLegendreTracking: changed algorithm for fillin quadtree nodes; bugfixes
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@15761 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/modules/trackFinderCDC/include/TrackFinderCDCLegendreTrackingModule.h
@@ -98,7 +98,7 @@ namespace Belle2 {
 
   private:
 
-    typedef TrackFindingCDC::QuadTreeTemplate<double, int, TrackFindingCDC::TrackHit> QuadTreeLegendre;
+    typedef TrackFindingCDC::QuadTreeTemplate<int, double, TrackFindingCDC::TrackHit> QuadTreeLegendre;
 
     std::string m_cdcHitsColName; /**< Input digitized hits collection name (output of CDCDigitizer module) */
     std::string m_gfTrackCandsColName; /**< Output genfit track candidates collection name*/
